:recycle: refactor: 패키지 이동 및 클래스 이름 변경
기존 별도로 분리 되어 있던 숙소 검색 기능 관련 클래스와 패키지를 숙소 관련 패키지 내부로 이동
