+++ airdnb/src/main/java/com/example/airdnb/controller/AccommodationController.java
@@ -1,11 +1,12 @@
 package com.example.airdnb.controller;
 
 import com.example.airdnb.domain.accommodation.Accommodation;
+import com.example.airdnb.domain.accommodation.search.AccommodationSearchCond;
 import com.example.airdnb.domain.user.UserDetail;
 import com.example.airdnb.dto.accommodation.AccommodationCreationRequest;
-import com.example.airdnb.dto.accommodation.AccommodationResponse;
 import com.example.airdnb.dto.review.ReviewCreateRequest;
 import com.example.airdnb.dto.review.ReviewResponse;
+import com.example.airdnb.dto.search.AccommodationSearchCondRequest;
 import com.example.airdnb.service.AccommodationService;
 import com.example.airdnb.service.ReviewService;
 import jakarta.validation.Valid;
@@ -17,6 +18,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -33,8 +35,11 @@ public class AccommodationController {
     private final ReviewService reviewService;
 
     @GetMapping
-    public List<AccommodationResponse> getAccommodationList() {
-        return accommodationService.getAccommodationList();
+    public List<Accommodation> search(@ModelAttribute @Valid AccommodationSearchCondRequest accommodationSearchCondRequest) {
+
+        AccommodationSearchCond accommodationSearchCond = accommodationSearchCondRequest.toEntity();
+        log.info(accommodationSearchCond.toString());
+        return accommodationService.searchWithCondition(accommodationSearchCond);
     }
 
     @PostMapping
