+++ airdnb/src/main/java/com/example/airdnb/domain/accommodation/search/StayPeriod.java
@@ -1,4 +1,4 @@
-package com.example.airdnb.domain.search;
+package com.example.airdnb.domain.accommodation.search;
 
 import java.time.LocalDate;
 import lombok.Getter;
