+++ airdnb/src/main/java/com/example/airdnb/repository/AccommodationRepositoryCustom.java
@@ -1,10 +1,10 @@
 package com.example.airdnb.repository;
 
 import com.example.airdnb.domain.accommodation.Accommodation;
-import com.example.airdnb.domain.search.SearchCondition;
+import com.example.airdnb.domain.accommodation.search.AccommodationSearchCond;
 import java.util.List;
 
 public interface AccommodationRepositoryCustom {
 
-    List<Accommodation> search(SearchCondition searchCondition);
+    List<Accommodation> search(AccommodationSearchCond accommodationSearchCond);
 }
