+++ airdnb/src/main/java/com/example/airdnb/service/AccommodationService.java
@@ -1,10 +1,10 @@
 package com.example.airdnb.service;
 
 import com.example.airdnb.domain.accommodation.Accommodation;
+import com.example.airdnb.domain.accommodation.search.AccommodationSearchCond;
 import com.example.airdnb.domain.user.User;
 import com.example.airdnb.domain.user.User.Role;
 import com.example.airdnb.dto.accommodation.AccommodationCreationRequest;
-import com.example.airdnb.dto.accommodation.AccommodationResponse;
 import com.example.airdnb.repository.AccommodationRepository;
 import com.example.airdnb.repository.UserRepository;
 import java.util.List;
@@ -18,13 +18,8 @@ public class AccommodationService {
     private final AccommodationRepository accommodationRepository;
     private final UserRepository userRepository;
 
-    public List<AccommodationResponse> getAccommodationList() { // 검색 조건에 따른 결과를 가져오도록 변경 해야함
-
-        List<Accommodation> accommodations = accommodationRepository.findAll();
-
-        return accommodations.stream()
-                .map(AccommodationResponse::of)
-                .toList();
+    public List<Accommodation> searchWithCondition(AccommodationSearchCond searchCond) {
+        return accommodationRepository.search(searchCond);
     }
 
     public Accommodation createNewAccommodation(AccommodationCreationRequest request) {
