--- airdnb/src/main/java/com/example/airdnb/service/SearchService.java
@@ -1,12 +1,0 @@
-package com.example.airdnb.service;
-
-import com.example.airdnb.domain.search.SearchCondition;
-import org.springframework.stereotype.Service;
-
-@Service
-public class SearchService {
-
-    public void searchWithConditions(SearchCondition searchCondition) {
-
-    }
-}
