+++ airdnb/src/test/java/com/example/airdnb/domain/search/PriceRangeTest.java
@@ -2,7 +2,9 @@ package com.example.airdnb.domain.search;
 
 import static org.assertj.core.api.Assertions.*;
 
-import com.example.airdnb.dto.search.SearchConditionRequest;
+import com.example.airdnb.domain.accommodation.search.PriceRange;
+import com.example.airdnb.domain.accommodation.search.AccommodationSearchCond;
+import com.example.airdnb.dto.search.AccommodationSearchCondRequest;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -28,14 +30,14 @@ public class PriceRangeTest {
     @Test
     @DisplayName("검색조건에서 PriceRange가 null 일 경우 지정된 기본값으로 객체가 생성되어야 한다\"")
     void priceRangeDefaultValueTest() {
-        SearchConditionRequest searchConditionRequest = new SearchConditionRequest(
-                StayPeriod.of(LocalDate.now(), LocalDate.now().plusDays(2)),
-                null,
+        AccommodationSearchCondRequest accommodationSearchCondRequest = new AccommodationSearchCondRequest(
+                LocalDate.now(), LocalDate.now().plusDays(2),
+                null, null,
                 2
         );
 
-        SearchCondition searchCondition = searchConditionRequest.toEntity();
-        PriceRange priceRange = searchCondition.priceRange();
+        AccommodationSearchCond accommodationSearchCond = accommodationSearchCondRequest.toEntity();
+        PriceRange priceRange = accommodationSearchCond.priceRange();
 
         assertThat(priceRange.getMinPrice()).isEqualTo(PriceRange.DEFAULT_MIN_VALUE);
         assertThat(priceRange.getMaxPrice()).isEqualTo(PriceRange.DEFAULT_MAX_VALUE);
